pyproject.toml
[tool.poetry]
name = "PDaltagent"
version = "0.5.0"
description = ""
authors = ["Martin Stone <martin@pagerduty.com>"]
exclude = ["pdaltagent/ui"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2"
celery = "^5"
flask = "^2"
gunicorn = "^20"
pymongo = "^4.2.0"
validators = "^0.20.0"
func-timeout = "^4.3.5"
croniter = "^1.3.8"
flask-wtf = "^1.2.1"
mongoengine = "^0.28.2"
flask-security-too = "^5.4.3"
bcrypt = "^4.0.1"
bleach = "^6.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7"

[tool.poetry.scripts]
pdagentd = 'pdaltagent.tasks:consume'
pdpollerd = 'pdaltagent.tasks:poll'
pd-send = 'pdaltagent.pdsend:main'
pdaltagentui = 'pdaltagent.api.api:app'

[build-system]
requires = ["poetry>=1.1.14"]
build-backend = "poetry.masonry.api"
